Increase the maximum time for a temporary role duration to 576000 minutes (400 days).
Note: The storage for events is bounded by the number of role assignments,
regardless of how long the events are stored. Therefore, there cannot be
any DoS vulnerability created by this PR that doesn't already exist as a
consequence of storing the role assignments.

But I guess it's pointless saying that because it would be an
"Internet argument" 🙄.

Signed-off-by: Daira Emma Hopwood <daira@jacaranda.org>